+++ app/apis/curriculums.py
@@ -1,12 +1,13 @@
 from logging import getLogger
 from typing import Annotated
+import json
 
 from database.database import get_db
 from fastapi import APIRouter, Depends, HTTPException, Path, Query
 from sqlalchemy.orm import Session
 from starlette import status
 
-from schemas.curriculums import ReviewsResponseBody, DetailResponseBody
+from schemas.curriculums import ReviewsResponseBody, DetailResponseBody,RequestBody, ResponseBody
 from cruds import curriculums as curriculums_crud
 
 logger = getLogger("uvicorn.app")
@@ -101,3 +102,74 @@ async def find_curriculum_details(db: DbDependency, curriculum_id: int = Path(gt
     if not info:
         raise HTTPException(status_code=404, detail="Curriculum not found.")
     return info
+
+@router.post("/{curriculum_id}/questions", response_model=ResponseBody, status_code=status.HTTP_201_CREATED)
+async def create_question(db: DbDependency, param:RequestBody, curriculum_id: int = Path(gt=0)):
+    """
+    質問投稿作成取得
+
+    Parameter
+    -----------------------
+    curriculum_id: int
+        詳細を取得したいカリキュラムのID
+    dict
+        user_id: int
+            ユーザーのID
+        title: str
+            質問のタイトル
+        content: str 
+            質問の内容
+        media_content: str
+            関連するメディアコンテンツの情報
+            url: str
+                メディアコンテンツのURL
+
+    Returns
+    -----------------------
+    dict
+        question_id: int
+            質問のID
+        curriculum_id: int
+            カリキュラムのID
+        user_id: int
+            ユーザーのID
+        title: str
+            質問のタイトル
+        content: str 
+            質問の内容
+        media_content: str 
+            関連するメディアコンテンツの情報
+    """
+    found_curriculum = curriculums_crud.find_curriculum(db, curriculum_id)
+
+    if not found_curriculum:
+        raise HTTPException(status_code=404,detail="Curriculum not found.")
+
+    di = {
+            "url": param.media_content.url
+        }
+    media_json = json.dumps(di)
+
+    try:
+        new_question = curriculums_crud.create_question(db, param.user_id, param.title, param.content, media_json, curriculum_id)
+        db.commit()
+        
+        re_di = {
+            "question_id": new_question.id,
+            "curriculum_id": new_question.curriculum_id,
+            "user_id": new_question.user_id,
+            "title": new_question.title,
+            "content": new_question.content,
+            "media_content": [
+                json.loads(new_question.media_content)
+            ]
+        }
+
+        return re_di
+
+    except Exception as e:
+        logger.error(str(e)) 
+        db.rollback()
+        raise HTTPException(status_code=400, detail="Invalid input data.")     
+
+
